Fixes for retrieve annotations from biomaRt
This deals with 2 issues: (1) In the latest version of biomaRt, the entrez identifier attribute has changed from "entrezgene" to "entrezgene_id", (2) The biomaRt host is occasionally down.

To fix (1) the function fetchAnnotations() will now auto-detect the appropriate column for the entrez identifiers from the results of a call to listAttributes(). To fix (2) it is now possible to specify one of the alternate ensembl hosts (e.g. useast.ensembl.org or uswest.ensembl.org) when calling fetchAnnotations().